Update login flow and integration with client
